back-end/requirements.txt
flask
flask-sqlalchemy
flask-jwt-extended
flask-cors
pymysql
cryptography
passlib[bcrypt]
pydantic
email-validator
python-dotenv